+++ packages/button/src/index.vue
@@ -105,7 +105,7 @@ export default defineComponent({
 
 <style>
 .btn {
-  @apply border transition px-4 py-2 rounded font-medium text-sm
+  @apply border border-solid cursor-pointer transition px-4 py-2 rounded font-medium text-sm
   focus:outline-none;
 }