+++ packages/dropdown/index.ts
@@ -1,0 +1,8 @@
+import { App } from 'vue'
+import Dropdown from './src/index.vue'
+
+Dropdown.install = (app: App): void => {
+  app.component(Dropdown.name, Dropdown)
+}
+
+export default Dropdown