+++ src/docs/en/button.vue
@@ -10,42 +10,42 @@
 
   <h2>Basic Usage</h2>
   <div class="demo">
-    <i-button class="mr-2 mb-3" type="default">Default</i-button>
-    <i-button class="mr-2 mb-3" type="secondary">Secondary</i-button>
-    <i-button class="mr-2 mb-3" type="primary">Primary</i-button>
-    <i-button class="mr-2 mb-3" type="success">Success</i-button>
-    <i-button class="mr-2 mb-3" type="info">Info</i-button>
-    <i-button class="mr-2 mb-3" type="warning">Warning</i-button>
-    <i-button class="mr-2 mb-3" type="danger">Danger</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="default">Default</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="secondary">Secondary</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="primary">Primary</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="success">Success</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="info">Info</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="warning">Warning</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="danger">Danger</i-button>
 
-    <hr class="mb-3">
-    <i-button plain class="mr-2 mb-3" type="default">Default</i-button>
-    <i-button plain class="mr-2 mb-3" type="secondary">Secondary</i-button>
-    <i-button plain class="mr-2 mb-3" type="primary">Primary</i-button>
-    <i-button plain class="mr-2 mb-3" type="success">Success</i-button>
-    <i-button plain class="mr-2 mb-3" type="info">Info</i-button>
-    <i-button plain class="mr-2 mb-3" type="warning">Warning</i-button>
-    <i-button plain class="mr-2 mb-3" type="danger">Danger</i-button>
+    <i-divider />
+    <i-button plain class="mr-2 mb-3 md:mb-0" type="default">Default</i-button>
+    <i-button plain class="mr-2 mb-3 md:mb-0" type="secondary">Secondary</i-button>
+    <i-button plain class="mr-2 mb-3 md:mb-0" type="primary">Primary</i-button>
+    <i-button plain class="mr-2 mb-3 md:mb-0" type="success">Success</i-button>
+    <i-button plain class="mr-2 mb-3 md:mb-0" type="info">Info</i-button>
+    <i-button plain class="mr-2 mb-3 md:mb-0" type="warning">Warning</i-button>
+    <i-button plain class="mr-2 mb-3 md:mb-0" type="danger">Danger</i-button>
 
-    <hr class="mb-3">
-    <i-button round class="mr-2 mb-3" type="default">Default</i-button>
-    <i-button round class="mr-2 mb-3" type="secondary">Secondary</i-button>
-    <i-button round class="mr-2 mb-3" type="primary">Primary</i-button>
-    <i-button round class="mr-2 mb-3" type="success">Success</i-button>
-    <i-button round class="mr-2 mb-3" type="info">Info</i-button>
-    <i-button round class="mr-2 mb-3" type="warning">Warning</i-button>
-    <i-button round class="mr-2 mb-3" type="danger">Danger</i-button>
+    <i-divider />
+    <i-button round class="mr-2 mb-3 md:mb-0" type="default">Default</i-button>
+    <i-button round class="mr-2 mb-3 md:mb-0" type="secondary">Secondary</i-button>
+    <i-button round class="mr-2 mb-3 md:mb-0" type="primary">Primary</i-button>
+    <i-button round class="mr-2 mb-3 md:mb-0" type="success">Success</i-button>
+    <i-button round class="mr-2 mb-3 md:mb-0" type="info">Info</i-button>
+    <i-button round class="mr-2 mb-3 md:mb-0" type="warning">Warning</i-button>
+    <i-button round class="mr-2 mb-3 md:mb-0" type="danger">Danger</i-button>
 
-    <hr class="mb-3">
-    <i-button flat class="mr-2 mb-3" type="default">Default</i-button>
-    <i-button flat class="mr-2 mb-3" type="secondary">Secondary</i-button>
-    <i-button flat class="mr-2 mb-3" type="primary">Primary</i-button>
-    <i-button flat class="mr-2 mb-3" type="success">Success</i-button>
-    <i-button flat class="mr-2 mb-3" type="info">Info</i-button>
-    <i-button flat class="mr-2 mb-3" type="warning">Warning</i-button>
-    <i-button flat class="mr-2 mb-3" type="danger">Danger</i-button>
+    <i-divider />
+    <i-button flat class="mr-2 mb-3 md:mb-0" type="default">Default</i-button>
+    <i-button flat class="mr-2 mb-3 md:mb-0" type="secondary">Secondary</i-button>
+    <i-button flat class="mr-2 mb-3 md:mb-0" type="primary">Primary</i-button>
+    <i-button flat class="mr-2 mb-3 md:mb-0" type="success">Success</i-button>
+    <i-button flat class="mr-2 mb-3 md:mb-0" type="info">Info</i-button>
+    <i-button flat class="mr-2 mb-3 md:mb-0" type="warning">Warning</i-button>
+    <i-button flat class="mr-2 mb-3 md:mb-0" type="danger">Danger</i-button>
 
-    <hr class="mb-3">
+    <i-divider />
     <i-button block type="primary">Block Button</i-button>
   </div>
   <pre class="ispa-code language-html"><code>{{ code3 }}</code></pre>
@@ -54,22 +54,22 @@
   <h2>Disabled Button</h2>
   <p>The <code>disabled</code> attribute determines if the button is disabled.</p>
   <div class="demo">
-    <i-button disabled class="mr-2 mb-3" type="default">Default</i-button>
-    <i-button disabled class="mr-2 mb-3" type="secondary">Secondary</i-button>
-    <i-button disabled class="mr-2 mb-3" type="primary">Primary</i-button>
-    <i-button disabled class="mr-2 mb-3" type="success">Success</i-button>
-    <i-button disabled class="mr-2 mb-3" type="info">Info</i-button>
-    <i-button disabled class="mr-2 mb-3" type="warning">Warning</i-button>
-    <i-button disabled class="mr-2 mb-3" type="danger">Danger</i-button>
+    <i-button disabled class="mr-2 mb-3 md:mb-0" type="default">Default</i-button>
+    <i-button disabled class="mr-2 mb-3 md:mb-0" type="secondary">Secondary</i-button>
+    <i-button disabled class="mr-2 mb-3 md:mb-0" type="primary">Primary</i-button>
+    <i-button disabled class="mr-2 mb-3 md:mb-0" type="success">Success</i-button>
+    <i-button disabled class="mr-2 mb-3 md:mb-0" type="info">Info</i-button>
+    <i-button disabled class="mr-2 mb-3 md:mb-0" type="warning">Warning</i-button>
+    <i-button disabled class="mr-2 mb-3 md:mb-0" type="danger">Danger</i-button>
 
-    <hr class="mb-3">
-    <i-button disabled plain class="mr-2 mb-3" type="default">Default</i-button>
-    <i-button disabled plain class="mr-2 mb-3" type="secondary">Secondary</i-button>
-    <i-button disabled plain class="mr-2 mb-3" type="primary">Primary</i-button>
-    <i-button disabled plain class="mr-2 mb-3" type="success">Success</i-button>
-    <i-button disabled plain class="mr-2 mb-3" type="info">Info</i-button>
-    <i-button disabled plain class="mr-2 mb-3" type="warning">Warning</i-button>
-    <i-button disabled plain class="mr-2 mb-3" type="danger">Danger</i-button>
+    <i-divider />
+    <i-button disabled plain class="mr-2 mb-3 md:mb-0" type="default">Default</i-button>
+    <i-button disabled plain class="mr-2 mb-3 md:mb-0" type="secondary">Secondary</i-button>
+    <i-button disabled plain class="mr-2 mb-3 md:mb-0" type="primary">Primary</i-button>
+    <i-button disabled plain class="mr-2 mb-3 md:mb-0" type="success">Success</i-button>
+    <i-button disabled plain class="mr-2 mb-3 md:mb-0" type="info">Info</i-button>
+    <i-button disabled plain class="mr-2 mb-3 md:mb-0" type="warning">Warning</i-button>
+    <i-button disabled plain class="mr-2 mb-3 md:mb-0" type="danger">Danger</i-button>
   </div>
   <pre class="ispa-code language-html"><code>{{ `<i-button disabled type="primary">Primary</i-button>` }}</code></pre>
 
@@ -77,18 +77,18 @@
   <h2>Loading Button</h2>
   <p>Display a loading state button.</p>
   <div class="demo">
-    <i-button loading class="mr-2 mb-3" type="primary">Loading Button</i-button>
+    <i-button loading class="mr-2 mb-3 md:mb-0" type="primary">Loading Button</i-button>
 
-    <hr class="mb-3">
+    <i-divider />
 
     <div class="block mb-2">
       <label for="loading_state">
         <input id="loading_state" class="mr-1" type="checkbox" :checked="loadingState" v-model="loadingState"> Loading: <b>{{ loadingState ? 'true' : 'false' }}</b>
       </label>
     </div>
-    <i-button :loading="loadingState" class="mr-2 mb-3" type="primary">Loading Primary</i-button>
-    <i-button :loading="loadingState" class="mr-2 mb-3" icon="fa fa-trash-alt" type="danger">Loading with Icon</i-button>
-    <i-button :loading="loadingState" class="mr-2 mb-3" icon="fa fa-trash-alt" type="danger" icon-align="right">Loading with Icon Right</i-button>
+    <i-button :loading="loadingState" class="mr-2 mb-3 md:mb-0" type="primary">Loading Primary</i-button>
+    <i-button :loading="loadingState" class="mr-2 mb-3 md:mb-0" icon="fa fa-trash-alt" type="danger">Loading with Icon</i-button>
+    <i-button :loading="loadingState" class="mr-2 mb-3 md:mb-0" icon="fa fa-trash-alt" type="danger" icon-align="right">Loading with Icon Right</i-button>
   </div>
   <pre class="ispa-code language-html"><code>{{ `<i-button loading type="primary">Loading Button</i-button>` }}</code></pre>
 
@@ -96,13 +96,13 @@
   <h2>Icons & Align</h2>
   <p>Use icons to add more meaning to Button. You can use icon alone to save some space, or use it with text.</p>
   <div class="demo">
-    <i-button class="mr-2 mb-3" type="primary" icon="fa fa-edit" />
-    <i-button class="mr-2 mb-3" type="danger" icon="fa fa-trash-alt" />
-    <i-button class="mr-2 mb-3" type="info" icon="fa fa-search">Search</i-button>
-    <i-button class="mr-2 mb-3" type="success" icon="fa fa-cloud-upload" icon-align="right">Upload</i-button>
-    <i-button loading class="mr-2 mb-3" type="danger" icon="fa fa-save" icon-align="right" />
-    <i-button class="mr-2 mb-3" size="sm" type="danger" icon="fa fa-save" icon-align="right" />
-    <i-button class="mr-2 mb-3" size="xs" type="danger" icon="fa fa-save" icon-align="right" />
+    <i-button class="mr-2 mb-3 md:mb-0" type="primary" icon="fa fa-edit" />
+    <i-button class="mr-2 mb-3 md:mb-0" type="danger" icon="fa fa-trash-alt" />
+    <i-button class="mr-2 mb-3 md:mb-0" type="info" icon="fa fa-search">Search</i-button>
+    <i-button class="mr-2 mb-3 md:mb-0" type="success" icon="fa fa-cloud-upload" icon-align="right">Upload</i-button>
+    <i-button loading class="mr-2 mb-3 md:mb-0" type="danger" icon="fa fa-save" icon-align="right" />
+    <i-button class="mr-2 mb-3 md:mb-0" size="sm" type="danger" icon="fa fa-save" icon-align="right" />
+    <i-button class="mr-2 mb-3 md:mb-0" size="xs" type="danger" icon="fa fa-save" icon-align="right" />
   </div>
   <pre class="ispa-code language-html"><code>{{ `<i-button type="danger" icon="fa fa-trash-alt" />
 <i-button type="info" icon="fa fa-search" icon-align="right">Search</i-button>` }}</code></pre>
@@ -111,11 +111,11 @@
   <h2>Sizes</h2>
   <p>Besides default size, Button component provides four additional sizes for you to choose among different scenarios.</p>
   <div class="demo">
-    <i-button size="xl" class="mr-2 mb-3" type="primary">Size xl</i-button>
-    <i-button size="lg" class="mr-2 mb-3" type="primary">Size lg</i-button>
-    <i-button size="md" class="mr-2 mb-3" type="primary">Size md (default)</i-button>
-    <i-button size="sm" class="mr-2 mb-3" type="primary">Size sm</i-button>
-    <i-button size="xs" class="mr-2 mb-3" type="primary">Size xs</i-button>
+    <i-button size="xl" class="mr-2 mb-3 md:mb-0" type="primary">Size xl</i-button>
+    <i-button size="lg" class="mr-2 mb-3 md:mb-0" type="primary">Size lg</i-button>
+    <i-button size="md" class="mr-2 mb-3 md:mb-0" type="primary">Size md (default)</i-button>
+    <i-button size="sm" class="mr-2 mb-3 md:mb-0" type="primary">Size sm</i-button>
+    <i-button size="xs" class="mr-2 mb-3 md:mb-0" type="primary">Size xs</i-button>
   </div>
   <pre class="ispa-code language-html"><code>{{ code4 }}</code></pre>