+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+#### --- Version 0.7.0
+This version update:
+* Added new element `divider`, `dropdown`
+* Fixed style of `alert`, `button`, `switch` (no need to required TailwindCss)
+
 #### --- Version 0.5.0
 This version update:
 * Added new element `switch`